+++ src/wren_value.c
@@ -369,8 +369,8 @@ static uint32_t hashObject(Obj* object)
       // towards the prefix or suffix of the string. So sample up to eight
       // characters spread throughout the string.
       // TODO: Tune this.
-      uint32_t step = 1 + 7 / string->length;
-      for (uint32_t i = 0; i < string->length; i += step)
+      int step = 1 + 7 / string->length;
+      for (int i = 0; i < string->length; i += step)
       {
         hash ^= string->value[i];
         hash *= 16777619;
@@ -596,7 +596,7 @@ Value wrenNewRange(WrenVM* vm, double from, double to, bool isInclusive)
   return OBJ_VAL(range);
 }
 
-Value wrenNewString(WrenVM* vm, const char* text, size_t length)
+Value wrenNewString(WrenVM* vm, const char* text, int length)
 {
   // Allow NULL if the string is empty since byte buffers don't allocate any
   // characters for a zero-length string.
@@ -613,19 +613,19 @@ Value wrenNewString(WrenVM* vm, const char* text, size_t length)
   return OBJ_VAL(string);
 }
 
-Value wrenNewUninitializedString(WrenVM* vm, size_t length)
+Value wrenNewUninitializedString(WrenVM* vm, int length)
 {
   ObjString* string = ALLOCATE_FLEX(vm, ObjString, length + 1);
   initObj(vm, &string->obj, OBJ_STRING, vm->stringClass);
-  string->length = (int)length;
+  string->length = length;
 
   return OBJ_VAL(string);
 }
 
 ObjString* wrenStringConcat(WrenVM* vm, const char* left, const char* right)
 {
-  size_t leftLength = strlen(left);
-  size_t rightLength = strlen(right);
+  int leftLength = strlen(left);
+  int rightLength = strlen(right);
 
   Value value = wrenNewUninitializedString(vm, leftLength + rightLength);
   ObjString* string = AS_STRING(value);
